# go/src/api/swagger/swagger.json
{"swagger":"2.0","info":{"title":"The virtual wine cellar","description":"A simple goa service","version":""},"host":"localhost:8080","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/bottles/{bottleID}":{"get":{"tags":["bottle"],"summary":"show bottle","description":"Get bottle by id","operationId":"bottle#show","produces":["application/vnd.goa.example.bottle+json"],"parameters":[{"name":"bottleID","in":"path","description":"Bottle ID","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleBottle"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}},"/swagger.yaml":{"get":{"summary":"Download swagger/swagger.yaml","operationId":"swagger#/swagger.yaml","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"GoaExampleBottle":{"title":"Mediatype identifier: application/vnd.goa.example.bottle+json; view=default","type":"object","properties":{"href":{"type":"string","description":"API href for making requests on the bottle","example":"Voluptates non excepturi."},"id":{"type":"integer","description":"Unique bottle ID","example":6996573674690312810,"format":"int64"},"name":{"type":"string","description":"Name of wine","example":"Tenetur et saepe ratione debitis enim sapiente."}},"description":"A bottle of wine (default view)","example":{"href":"Voluptates non excepturi.","id":6996573674690312810,"name":"Tenetur et saepe ratione debitis enim sapiente."},"required":["id","href","name"]}},"responses":{"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleBottle"}}}}